+++ autd_firmware_writer.ps1
@@ -3,14 +3,14 @@
 # Created Date: 14/02/2020
 # Author: Shun Suzuki
 # -----
-# Last Modified: 14/12/2023
+# Last Modified: 26/12/2023
 # Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 # -----
 # Copyright (c) 2021 Shun Suzuki. All rights reserved.
 # 
 
 Param(
-    [string]$version = "4.1.2",
+    [string]$version = "5.0.0",
     [string]$vivado_dir = "NULL"
 )